Change to groupByDevice execution logic: processing filter by device separately
