+++ server/src/main/java/org/apache/iotdb/db/qp/strategy/PhysicalGenerator.java
@@ -24,6 +24,7 @@ import org.apache.iotdb.db.exception.path.PathException;
 import org.apache.iotdb.db.exception.query.LogicalOperatorException;
 import org.apache.iotdb.db.exception.query.LogicalOptimizeException;
 import org.apache.iotdb.db.exception.query.QueryProcessException;
+import org.apache.iotdb.db.metadata.MManager;
 import org.apache.iotdb.db.qp.constant.SQLConstant;
 import org.apache.iotdb.db.qp.executor.IQueryProcessExecutor;
 import org.apache.iotdb.db.qp.logical.Operator;
@@ -313,25 +314,93 @@ public class PhysicalGenerator {
       queryPlan.setDataTypeConsistencyChecker(dataTypeConsistencyChecker);
       queryPlan.setPaths(paths);
       queryPlan.setDeduplicatedPaths(paths);
+
+      // get device to filter map
+      FilterOperator filterOperator = queryOperator.getFilterOperator();
+
+      if(filterOperator != null){
+        queryPlan.setDeviceToFilterMap(concatFilterByDivice(prefixPaths, filterOperator));
+      }
     } else {
       List<Path> paths = queryOperator.getSelectedPaths();
       queryPlan.setPaths(paths);
+
+      // transform filter operator to expression
+      FilterOperator filterOperator = queryOperator.getFilterOperator();
+
+      if (filterOperator != null) {
+        IExpression expression = filterOperator.transformToExpression(executor);
+        queryPlan.setExpression(expression);
+      }
     }
     generateDataTypes(queryPlan);
     deduplicate(queryPlan);
-
-    // transform filter operator to expression
-    FilterOperator filterOperator = queryOperator.getFilterOperator();
-
-    if (filterOperator != null) {
-      IExpression expression = filterOperator.transformToExpression(executor);
-      queryPlan.setExpression(expression);
-    }
 
     queryPlan.setRowLimit(queryOperator.getRowLimit());
     queryPlan.setRowOffset(queryOperator.getRowOffset());
 
     return queryPlan;
+  }
+
+  // e.g. translate "select * from root.ln.d1, root.ln.d2 where s1 < 20 AND s2 > 10" to
+  // [root.ln.d1 -> root.ln.d1.s1 < 20 AND root.ln.d1.s2 > 10,
+  //  root.ln.d2 -> root.ln.d2.s1 < 20 AND root.ln.d2.s2 > 10)]
+  private Map<String, IExpression> concatFilterByDivice(List<Path> fromPaths, FilterOperator operator)
+          throws QueryProcessException {
+    Map<String, IExpression> deviceToFilterMap = new HashMap<>();
+    // remove stars in fromPaths and get deviceId with deduplication
+    List<String> noStarDevices = removeStarsInDeviceWithUnique(fromPaths);
+    for (int i = 0; i < noStarDevices.size(); i++) {
+      FilterOperator newOperator = operator.clone();
+      newOperator = concatFilterPath(noStarDevices.get(i), newOperator);
+
+      deviceToFilterMap.put(noStarDevices.get(i), newOperator.transformToExpression(executor));
+    }
+
+    return deviceToFilterMap;
+  }
+
+  private List<String> removeStarsInDeviceWithUnique(List<Path> paths)
+          throws LogicalOptimizeException {
+    List<String> retDevices;
+    Set<String> deviceSet = new LinkedHashSet<>();
+    try {
+      for (Path path : paths) {
+        List<String> tempDS;
+        tempDS = MManager.getInstance().getDevices(path.getFullPath());
+
+        for (String subDevice : tempDS) {
+          if (!deviceSet.contains(subDevice)) {
+            deviceSet.add(subDevice);
+          }
+        }
+      }
+      retDevices = new ArrayList<>(deviceSet);
+    } catch (PathException e) {
+      throw new LogicalOptimizeException("error when remove star: " + e.getMessage());
+    }
+    return retDevices;
+  }
+
+  private FilterOperator concatFilterPath(String prefix, FilterOperator operator) {
+    if(!operator.isLeaf()){
+      for (FilterOperator child : operator.getChildren()) {
+        concatFilterPath(prefix, child);
+      }
+      return operator;
+    }
+    BasicFunctionOperator basicOperator = (BasicFunctionOperator) operator;
+    Path filterPath = basicOperator.getSinglePath();
+
+    // do nothing in the cases of "where time > 5" or "where root.d1.s1 > 5"
+    if (SQLConstant.isReservedPath(filterPath) || filterPath.startWith(SQLConstant.ROOT)) {
+      return operator;
+    }
+
+    Path concatPath = filterPath.addPrefixPath(filterPath, prefix);
+    basicOperator.setSinglePath(concatPath);
+
+    return basicOperator;
   }
 
   private void generateDataTypes(QueryPlan queryPlan) throws PathException {
